Fix return value for `mod` and `vm` when using globbing

+++ src/cmdlets/mod.ts
@@ -91,7 +91,7 @@ export class ModCmdLet extends CmdLet {
     });
     if (modules.length === 1) {
       const module = modules[0] as Module;
-      return new Var(module.base.toString());
+      return new Var(uint64(module.base.toString()));
     } else {
       return Var.ZERO;
     }
